Add notification rate limit sensor
