+++ Sources/App/Settings/Notifications/NotificationRateLimitViewController.swift
@@ -162,3 +162,48 @@ class NotificationRateLimitListViewController: HAFormViewController {
         row.updateCell()
     }
 }
+
+public extension RateLimitResponse.RateLimits {
+    func row(for keyPath: KeyPath<Self, Int>) -> BaseRow {
+        LabelRow {
+            $0.value = NumberFormatter.localizedString(
+                from: NSNumber(value: self[keyPath: keyPath]),
+                number: .none
+            )
+            $0.title = { () -> String in
+                switch keyPath {
+                case \.attempts:
+                    return L10n.SettingsDetails.Notifications.RateLimits.attempts
+                case \.successful:
+                    return L10n.SettingsDetails.Notifications.RateLimits.delivered
+                case \.errors:
+                    return L10n.SettingsDetails.Notifications.RateLimits.errors
+                case \.total:
+                    return L10n.SettingsDetails.Notifications.RateLimits.total
+                case \.maximum:
+                    return ""
+                default:
+                    fatalError("missing key: \(keyPath)")
+                }
+            }()
+        }
+    }
+
+    func row(for keyPath: KeyPath<Self, Date>) -> BaseRow {
+        LabelRow { row in
+            row.value = DateFormatter.localizedString(
+                from: self[keyPath: keyPath],
+                dateStyle: .none,
+                timeStyle: .medium
+            )
+
+            switch keyPath {
+            case \.resetsAt:
+                row.title = L10n.SettingsDetails.Notifications.RateLimits.resetsIn
+                row.tag = "resetsIn"
+            default:
+                fatalError("missing key: \(keyPath)")
+            }
+        }
+    }
+}
